New warnings for arm64.

+++ OHAttributedLabel/Source/NSAttributedString+Attributes.m
@@ -78,7 +78,7 @@ NSString* kOHLinkAttributeName = @"NSLinkAttributeName"; // Use the same value a
     {
         CFRange fitCFRange = CFRangeMake(0,0);
         sz = CTFramesetterSuggestFrameSizeWithConstraints(framesetter,CFRangeMake(0,0),NULL,maxSize,&fitCFRange);
-        sz = CGSizeMake( floorf(sz.width+1) , floorf(sz.height+1) ); // take 1pt of margin for security
+        sz = CGSizeMake( floorf((float)(sz.width+1)), floorf((float)(sz.height+1)) ); // take 1pt of margin for security
         
         if (maxLines > 0) {
             CGRect frameRect = (CGRect){.origin=CGPointZero, .size=maxSize};
@@ -382,7 +382,7 @@ static NSString* const kHelveticaNeueUI_Bold_Italic = @".HelveticaNeueUI-BoldIta
 }
 -(void)setCharacterSpacing:(CGFloat)chracterSpacing range:(NSRange)range
 {
-    [self addAttribute:(NSString *)kCTKernAttributeName value:[NSNumber numberWithFloat:chracterSpacing] range:range];
+    [self addAttribute:(NSString *)kCTKernAttributeName value:[NSNumber numberWithFloat:(float)chracterSpacing] range:range];
 }
 
 -(void)modifyParagraphStylesWithBlock:(void(^)(OHParagraphStyle* paragraphStyle))block
